+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Bug Fixes
 
  * Fix This annotation is not applicable to target 'interface'. Applicable targets: type usage error
+ * Fix expected value not correct error
 
 ### Features